python_app/requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
pytest>=7.4.0

